02 -Requisição DELETE para cancelar uma adoção pelo id do Pet

+++ src/main/java/com/adopet/apiadopet/services/PetService.java
@@ -58,8 +58,8 @@ public class PetService {
 
 	public void deletarPetPorId(Long id) {
 		var pet = petRepository.findById(id);
-		if (pet.isEmpty()) {
-			throw new ObjetoNaoEncontrado("Não foi possível deletar.Pet não encontrado");
+		if (pet.isEmpty() || pet.get().getAdotado() == true) {
+			throw new ObjetoNaoEncontrado("Não foi possível deletar.Pet não encontrado ou já foi adotado.");
 		}
 		petRepository.delete(pet.get());
 	}
